Fetch currency chart data and display it

+++ app/src/main/kotlin/com/makeuseof/cryptocurrency/data/model/ChartData.kt
@@ -4,6 +4,6 @@ import com.google.gson.annotations.SerializedName
 
 // Created by askar on 7/25/18.
 data class ChartData(
-        @SerializedName("price_usd") val usdChart: List<List<Long>>
+        @SerializedName("price_usd") val usdChart: List<List<Double>>
 ) {
 }
